updated readme, issue #48 fixed, updated yaml comments, updated dependencies
