+++ package.json
@@ -12,50 +12,54 @@
     "deploy": "webpack --mode=production --config webpack.prod.js && npx tailwindcss build src/components/tailwind.css -o dist/assets/tailwind.css.liquid && cleancss -o dist/assets/tailwind.min.css.liquid dist/assets/tailwind.css.liquid && shopify-themekit deploy --env=production"
   },
   "dependencies": {
-    "validator": "^13.1.1"
+    "validator": "^13.5.2"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.4",
-    "@babel/plugin-transform-runtime": "^7.12.1",
-    "@babel/preset-env": "^7.11.0",
-    "@shopify/eslint-plugin": "^39.0.3",
-    "@shopify/themekit": "^1.1.6",
-    "autoprefixer": "^10.0.4",
-    "axios": "^0.21.0",
+    "@babel/core": "^7.12.10",
+    "@babel/plugin-transform-runtime": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@shopify/eslint-plugin": "^39.0.4",
+    "@shopify/themekit": "^1.1.7",
+    "autoprefixer": "^10.2.3",
+    "axios": "^0.21.1",
     "babel-eslint": "^10.1.0",
-    "babel-loader": "^8.1.0",
+    "babel-loader": "^8.2.2",
     "babel-plugin-transform-class-properties": "^6.24.1",
-    "browser-sync": "^2.26.12",
+    "browser-sync": "^2.26.14",
     "clean-css-cli": "^4.3.0",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^6.3.0",
+    "copy-webpack-plugin": "6.x",
+    "core-js": "^3.8.3",
     "css-loader": "^5.0.1",
+    "debug": "^4.3.1",
     "eslint": "7.18.0",
     "eslint-config-prettier": "^7.2.0",
-    "eslint-webpack-plugin": "^2.4.0",
-    "file-loader": "^6.0.0",
+    "eslint-webpack-plugin": "^2.4.3",
+    "file-loader": "^6.2.0",
     "glob": "^7.1.6",
-    "html-webpack-plugin": "^4.3.0",
-    "liquidjs": "^9.16.1",
-    "mini-css-extract-plugin": "^1.3.1",
+    "graphql": "^15.5.0",
+    "html-webpack-plugin": "^4.5.1",
+    "liquidjs": "^9.20.1",
+    "mini-css-extract-plugin": "^1.3.4",
     "node-fetch": "^2.6.1",
     "node-sass": "^5.0.0",
-    "postcss": "^8.1.10",
-    "postcss-loader": "^4.0.4",
-    "prettier": "^2.1.2",
+    "postcss": "^8.2.4",
+    "postcss-loader": "^4.2.0",
+    "prettier": "^2.2.1",
     "raw-loader": "^4.0.2",
-    "sass": "^1.29.0",
-    "sass-loader": "^10.1.0",
+    "sass": "^1.32.5",
+    "sass-loader": "^10.1.1",
     "style-loader": "^2.0.0",
-    "tailwindcss": "^2.0.1",
+    "tailwindcss": "^2.0.2",
     "transform-class-properties": "^1.0.0-beta",
-    "webpack": "^5.4.0",
-    "webpack-bundle-analyzer": "^4.1.0",
-    "webpack-cli": "^4.2.0",
-    "webpack-dev-server": "^3.11.0",
+    "typescript": "^4.1.3",
+    "webpack": "^5.18.0",
+    "webpack-bundle-analyzer": "^4.4.0",
+    "webpack-cli": "^4.4.0",
+    "webpack-dev-server": "^3.11.2",
     "webpack-merge": "^5.7.3",
-    "webpack-shell-plugin-next": "^2.0.8",
+    "webpack-shell-plugin-next": "^2.1.1",
     "yaml": "^1.10.0",
-    "yargs": "^16.1.0"
+    "yargs": "^16.2.0"
   }
 }
