+++ webpack.dev.js
@@ -2,9 +2,7 @@ const { merge } = require('we' + 'bpack-merge');
 const glob = require('glob');
 const common = require('./webpack.common.js');
 const path = require('path');
-const CopyPlugin = require('copy-webpack-plugin');
 const WebpackShellPluginNext = require('webpack-shell-plugin-next');
-const { transformLiquid } = require('./shopify-dev-utils/transformLiquid');
 
 const port = 9000;
 const publicPath = `https://localhost:${port}/`;
@@ -99,35 +97,6 @@ module.exports = merge(common, {
                 blocking: true,
                 parallel: false,
             },
-        }),
-        new CopyPlugin({
-            patterns: [
-                {
-                    from: 'src/components/**/*.liquid',
-                    to: '[folder]/[name].[ext]',
-                    flatten: true,
-                    transformPath(targetPath, absolutePath) {
-                        const relativePath = path.join(__dirname, 'src/components');
-                        const diff = path.relative(relativePath, absolutePath);
-                        const targetFolder = diff.split(path.sep)[0];
-                        return path.join(targetFolder, path.basename(absolutePath));
-                    },
-                    transform: transformLiquid(publicPath),
-                },
-                {
-                    from: 'src/assets/**/*',
-                    to: 'assets/',
-                    flatten: true,
-                },
-                {
-                    from: 'src/config/*.json',
-                    to: 'config/[name].[ext]',
-                },
-                {
-                    from: 'src/locales/*.json',
-                    to: 'locales/[name].[ext]',
-                },
-            ],
         }),
     ],
     devServer: {
